requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.9
djangorestframework==3.12.4
idna==3.3
iniconfig==1.1.1
kombu==5.2.1
lxml==4.6.4
packaging==21.2
pluggy==1.0.0
prompt-toolkit==3.0.22
py==1.11.0
pyparsing==2.4.7
pytest==6.2.5
pytz==2021.3
redis==3.5.3
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
